+++ packages/kysely/migrate/migrations/2025-06-05T14:30:00_add_nonce_and_state_to_codes.ts
@@ -9,7 +9,7 @@ export async function up(db: Kysely<Database>): Promise<void> {
 
   await db.schema
     .alterTable("codes")
-    .addColumn("state", "varchar(1024)")
+    .addColumn("state", "varchar(2048)")
     .execute();
 }